etl/requirements.txt
SQLAlchemy>=2.0
PyMySQL>=1.1
pandas>=2.0
PyYAML>=6.0
python-dotenv>=1.0